move(logger): from internal/adapters to pkg fix(go.mod): change package name from github.com/Badsnus/cu-clubs-bot to github.com/Badsnus/cu-clubs-bot/bot
