+++ bot/cmd/bot/bot.go
@@ -3,9 +3,9 @@ package bot
 import (
 	"sync"
 
-	"github.com/Badsnus/cu-clubs-bot/internal/adapters/config"
-	"github.com/Badsnus/cu-clubs-bot/internal/adapters/logger"
-	"github.com/Badsnus/cu-clubs-bot/internal/adapters/logger/types"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/adapters/config"
+	"github.com/Badsnus/cu-clubs-bot/bot/pkg/logger"
+	"github.com/Badsnus/cu-clubs-bot/bot/pkg/logger/types"
 	"github.com/redis/go-redis/v9"
 	tele "gopkg.in/telebot.v3"
 	"gopkg.in/telebot.v3/layout"
