+++ bot/cmd/main.go
@@ -1,12 +1,12 @@
 package main
 
 import (
-	"github.com/Badsnus/cu-clubs-bot/cmd/bot"
-	setupBot "github.com/Badsnus/cu-clubs-bot/internal/adapters/controller/telegram/setup"
+	"github.com/Badsnus/cu-clubs-bot/bot/cmd/bot"
+	setupBot "github.com/Badsnus/cu-clubs-bot/bot/internal/adapters/controller/telegram/setup"
 
 	"log"
 
-	"github.com/Badsnus/cu-clubs-bot/internal/adapters/config"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/adapters/config"
 )
 
 func main() {
