+++ bot/internal/adapters/controller/telegram/handlers/admin.go
@@ -2,11 +2,11 @@ package handlers
 
 import (
 	"context"
-	"github.com/Badsnus/cu-clubs-bot/cmd/bot"
-	"github.com/Badsnus/cu-clubs-bot/internal/adapters/database/postgres"
-	"github.com/Badsnus/cu-clubs-bot/internal/adapters/database/redis"
-	"github.com/Badsnus/cu-clubs-bot/internal/domain/entity"
-	"github.com/Badsnus/cu-clubs-bot/internal/domain/service"
+	"github.com/Badsnus/cu-clubs-bot/bot/cmd/bot"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/adapters/database/postgres"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/adapters/database/redis"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/domain/entity"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/domain/service"
 	tele "gopkg.in/telebot.v3"
 	"gopkg.in/telebot.v3/layout"
 )
