+++ bot/internal/adapters/controller/telegram/handlers/onEvent.go
@@ -2,13 +2,13 @@ package handlers
 
 import (
 	"context"
-	"github.com/Badsnus/cu-clubs-bot/cmd/bot"
-	"github.com/Badsnus/cu-clubs-bot/internal/adapters/database/postgres"
-	"github.com/Badsnus/cu-clubs-bot/internal/adapters/database/redis"
-	states "github.com/Badsnus/cu-clubs-bot/internal/adapters/database/redis/state"
-	"github.com/Badsnus/cu-clubs-bot/internal/domain/common/errorz"
-	"github.com/Badsnus/cu-clubs-bot/internal/domain/entity"
-	"github.com/Badsnus/cu-clubs-bot/internal/domain/service"
+	"github.com/Badsnus/cu-clubs-bot/bot/cmd/bot"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/adapters/database/postgres"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/adapters/database/redis"
+	states "github.com/Badsnus/cu-clubs-bot/bot/internal/adapters/database/redis/state"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/domain/common/errorz"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/domain/entity"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/domain/service"
 	"github.com/spf13/viper"
 	tele "gopkg.in/telebot.v3"
 	"gopkg.in/telebot.v3/layout"
