+++ bot/internal/adapters/controller/telegram/handlers/user.go
@@ -4,12 +4,12 @@ import (
 	"context"
 	"time"
 
-	"github.com/Badsnus/cu-clubs-bot/cmd/bot"
-	"github.com/Badsnus/cu-clubs-bot/internal/adapters/database/postgres"
-	"github.com/Badsnus/cu-clubs-bot/internal/adapters/database/redis"
-	"github.com/Badsnus/cu-clubs-bot/internal/adapters/database/redis/state"
-	"github.com/Badsnus/cu-clubs-bot/internal/domain/entity"
-	"github.com/Badsnus/cu-clubs-bot/internal/domain/service"
+	"github.com/Badsnus/cu-clubs-bot/bot/cmd/bot"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/adapters/database/postgres"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/adapters/database/redis"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/adapters/database/redis/state"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/domain/entity"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/domain/service"
 	"github.com/spf13/viper"
 	tele "gopkg.in/telebot.v3"
 	"gopkg.in/telebot.v3/layout"
