+++ bot/internal/adapters/database/postgres/migrations.go
@@ -1,6 +1,6 @@
 package postgres
 
-import "github.com/Badsnus/cu-clubs-bot/internal/domain/entity"
+import "github.com/Badsnus/cu-clubs-bot/bot/internal/domain/entity"
 
 // Migrations is a list of all gorm migrations for the database.
 var Migrations = []interface{}{
