+++ bot/internal/adapters/database/postgres/user.go
@@ -3,7 +3,7 @@ package postgres
 import (
 	"context"
 
-	"github.com/Badsnus/cu-clubs-bot/internal/domain/entity"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/domain/entity"
 	"gorm.io/gorm"
 )
