+++ bot/internal/adapters/database/redis/state/state.go
@@ -1,7 +1,7 @@
 package state
 
 const (
-	WaitingForMailingContent            = "github.com/Badsnus/cu-clubs-bot"
+	WaitingForMailingContent            = "github.com/Badsnus/cu-clubs-bot/bot"
 	WaitingExternalUserFio              = "waiting_external_user_fio"
 	WaitingGrantUserFio                 = "waiting_grant_user_fio"
 	WaitingStudentEmail                 = "waiting_student_email"
