+++ bot/internal/adapters/database/redis/states.go
@@ -7,8 +7,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/Badsnus/cu-clubs-bot/cmd/bot"
-	"github.com/Badsnus/cu-clubs-bot/internal/domain/common/errorz"
+	"github.com/Badsnus/cu-clubs-bot/bot/cmd/bot"
+	"github.com/Badsnus/cu-clubs-bot/bot/internal/domain/common/errorz"
 	"github.com/redis/go-redis/v9"
 )
